Debug

+++ RaspberryPi/CodeWithROS/communication/src/uartparser.h
@@ -23,8 +23,8 @@ public:
 	~UARTparser();
 	void parseData(communication::msgStruct msg); // Wordt aangeroepen door receiveMsg()
 	void receiveMsg(communication::msgStruct& msg); // Callback van topic DriverParser1
-	void addPair(uint16_t, Parser*);
-	void removePair(uint16_t);
+	void addPair(uint32_t, Parser*);
+	void removePair(uint32_t);
 private:
 	void transmitInfo(communication::infoStruct info); // Publisht op topic ParserController1
 	std::map<uint16_t, Parser*> IDmap;
